New database model stage 1 (#18)
* New model approach stage 1

Changed how model.js works and thus updated a more generic solution for firebaselistener (second attempt due to new game logic)

* Condensing and simplifying stage 1 approach

Moving id back out and condensing where appropriate

* Fix adminUsers and remove groupmembership

* Quick hack of fakeserver

Race condition magically allows this to work on my computer, but remove it and it doesn't. There is work to be done, removing this.reader completely.

* missing playerid in notifcation

* missed private player in notifcation

* issue with notification and lives

for now the privatelife will have to have a separate call path to use the private exclusion problem path (private = obj not id)

* fix player removal

* admin users not available in time

Rely on the snap value instead, until we fill in the right information (can't because id = model).

* listen to game prop changes

* New model approach stage 1

Changed how model.js works and thus updated a more generic solution for firebaselistener (second attempt due to new game logic)

* Condensing and simplifying stage 1 approach

Moving id back out and condensing where appropriate

* Fix adminUsers and remove groupmembership

* Quick hack of fakeserver

Race condition magically allows this to work on my computer, but remove it and it doesn't. There is work to be done, removing this.reader completely.

* missing playerid in notifcation

* missed private player in notifcation

* issue with notification and lives

for now the privatelife will have to have a separate call path to use the private exclusion problem path (private = obj not id)

* fix player removal

* admin users not available in time

Rely on the snap value instead, until we fill in the right information (can't because id = model).

* listen to game prop changes

* update private life for infection