Faster loading and corrected rules.bolt (#41)
* Create firebaseobject counter

This is a quick patch which relies on a counter (yes, multi threaded race conditions are possible). There is a high probabilty this wont work on older browsers. Sorry.

* Messages speed boost

We can't control the message count, but we can control how many we listen to

* Don't add the last message twice

Bug here too, if there are 3 messages with the exact same timestamp then we could have a sorting problem

* Messages speed boost for realz

By leveraging the rules I found a bug and also boosted time to load by skipping the child_added for stale messages

* Simplify speed boost logic

Creates a min 500ms load time, could be variable but this number seems nice to ensure at least 1 call to firebase happens.

* Listen to 100 of the last message

* Fix privateLife rules

* Queuednotification updates