Fixed (kind of) a race condition in chat messages (#82)
This makes chat messages wait until a chat room exists. (also works around a seeming polymer bug, it gives us the wrong index in dom-repeats, when something's inserted in the beginning of the array)

Still a potential race condition: if we send a chat message and then leave a chat room (such as if we ever implement a "I have left the chat room" message), then it's possible that the leave will get through the server and into the firebase, before the chat message gets there. In this case, the chat message will be waiting forever in limbo.

It's a certain class of race condition we've opened ourselves up to now that we dont guarantee ordering.