Fix leaderboard sort stability
Currently, the leaderboard summary (top 3) on the Dashboard (web version) does not match the top 3 players on the Leaderboard tab.

This is because the template sorting functionality relies on the JS native sort, which does not perform stable sorting. Adding a unique key (`id`) to the sort's comparison function will produce a stable sorting order.

This new functionality will sort by `points` descending, with a secondary sort on player `id` ascending. This produces a leaderboard summary which will match up with the Leaderboard tab.