Show user name & avatar & timestamp in chat message
Shows the user name, avatar, & message timestamp in chat messages.
Also fixes chat room UI so that views are correctly sized and the recycler view
scrolls when it should now.
Also fixes the annoying keyboard toggle behavior, so the keyboard only opens
when opening a chat room and properly hides or stays hidden when exiting the
chat room.