Add functionality to create a chat room.
Hooks up the chat room FAB to support creating a chat room.
Fixed firebase functions to create a chat room with some default assumptions
(we'll need to update these for admin's chats)
Fixes the chat room list listeners so that view updates dynamically and
correctly when player is added to a chat room.